+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreePayPal
+  * Send the `edit_billing_agreement_jwt` sourced from the client token for the PayPal Edit FI beta flow
+
 ## 7.9.0 (2026-07-21)
 * BraintreeUIComponents
   * Fix minimum target version to iOS 16.0 to match all other modules